be/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
chromadb==0.4.22
sentence-transformers==2.2.2
numpy
pymongo
pydantic
pydantic-settings
PyJWT
requests
passlib[bcrypt]
python-jose[cryptography]
bcrypt
